Moved OptionState to core/options.py
